Added a load more importants from sklearn and also time for benchmarking. Made SVM and KMeans bits into a method each so I can call it as desired.
Had to sum the 4 sensor values in the sensor lists because it was making everything else break. It's not perfect and I'd like to have used each value to maybe correlate against the way they're moving and if it's harder or lighter between hinge and pendulum. Meh, too late now, that goes in future development.